Componentsの改修.
