+++ src/components/Sidebar/Filter/Filter.tsx
@@ -1,5 +1,5 @@
 import {FaInbox, FaRegCalendar, FaRegCalendarAlt} from 'react-icons/fa';
-import {NavLink} from 'react-router-dom';
+import {Link, LinkContent} from '@/components/UI';
 import styled from 'styled-components';
 
 export default function Filter() {
@@ -7,37 +7,37 @@ export default function Filter() {
     <>
       <StyledList>
         <StyledListItem>
-          <StyledNavLink to="/inbox">
-            <StyledIcon>
-              <FaInbox
-                size={24}
-                fill="#666"
+          <Link
+            to="/inbox"
+            children={
+              <LinkContent
+                icon={FaInbox}
+                label="インボックス"
               />
-            </StyledIcon>
-            <Label>インボックス</Label>
-          </StyledNavLink>
+            }
+          />
         </StyledListItem>
         <StyledListItem>
-          <StyledNavLink to="/today">
-            <StyledIcon>
-              <FaRegCalendar
-                size={24}
-                fill="#666"
+          <Link
+            to="/today"
+            children={
+              <LinkContent
+                icon={FaRegCalendar}
+                label="今日"
               />
-            </StyledIcon>
-            <Label>今日</Label>
-          </StyledNavLink>
+            }
+          />
         </StyledListItem>
         <StyledListItem>
-          <StyledNavLink to="/today">
-            <StyledIcon>
-              <FaRegCalendarAlt
-                size={24}
-                fill="#666"
+          <Link
+            to="/next"
+            children={
+              <LinkContent
+                icon={FaRegCalendarAlt}
+                label="近日予定"
               />
-            </StyledIcon>
-            <Label>近日予定</Label>
-          </StyledNavLink>
+            }
+          />
         </StyledListItem>
       </StyledList>
     </>
@@ -50,18 +50,4 @@ const StyledList = styled.ul`
 
 const StyledListItem = styled.li`
   padding: 3px 0px 3px 5px;
-`;
-
-const StyledNavLink = styled(NavLink)`
-  display: flex;
-  align-items: center;
-`;
-
-const StyledIcon = styled.span`
-  display: flex;
-  height: 14px;
-`;
-
-const Label = styled.span`
-  padding-left: 6px;
 `;
