+++ src/components/Task/index.ts
@@ -1,0 +1,3 @@
+export * from './AddButton';
+export * from './List';
+export * from './ListItem';
